An example commit of adding a test for a dts
